# docs/geo-rd-map-openapi.yaml
openapi: 3.0.3
info:
  title: GEO RD MAP API
  version: 2.6.1
  description: API geoespacial central de República Dominicana para integrar división territorial, selectores Provincia →
    Ciudad → Barrio, geocodificación, autocomplete, reverse geocoding, lugares y direcciones, GeoJSON/MVT, geocercas, routing
    y navegación. Las integraciones externas usan una API key privada por aplicación, permisos por scope, restricciones de
    origen y límites de consumo. Esta especificación documenta únicamente la API pública de integración; los endpoints /api/v1/admin
    forman parte del panel operativo y no son una API de consumo.
  contact:
    name: Administración de GEO RD MAP
servers:
- url: https://geo.ltd.do
  description: Producción
- url: http://localhost:8080
  description: Backend local
security:
- ApiKeyAuth: []
paths:
...
  /api/v1/addresses/validate:
    post:
      tags:
      - Geolocalización
      summary: Valida/enriquece una dirección contra los datos locales sincronizados
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddressValidationRequest'
            example:
              lat: 18.4861
              lng: -69.9312
              provinceCode: '01'
              cityId: 01:01:01
      responses:
        '200':
          description: Respuesta correcta
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AddressValidationResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '400':
          $ref: '#/components/responses/BadRequest'
      operationId: validateAddress
      security:
      - ApiKeyAuth: []
      x-required-scope: geo:read
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key privada de la aplicación. Las claves se crean en Integraciones API del panel administrativo. Cada
        clave puede limitarse por scopes, orígenes permitidos y rate limit.
...
  responses:
    BadRequest:
      description: Solicitud inválida.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: API key ausente, inválida o inactiva.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: La API key no posee el scope requerido o el origen no está autorizado.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    TooManyRequests:
      description: Límite global o por API key excedido.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
    ErrorResponse:
      type: object
      required:
      - valid
      - message
      properties:
        valid:
          type: boolean
          example: false
        message:
          type: string
          example: API Key inválida o inactiva
...
    Address:
      type: object
      required:
      - id
      - lat
      - lng
      properties:
        id:
          type: string
        number:
          type: string
        street:
          type: string
        unit:
          type: string
        postalCity:
          type: string
        postcode:
          type: string
        addressLevels:
          type: object
          additionalProperties: true
        lat:
          type: number
          format: double
        lng:
          type: number
          format: double
        distanceMeters:
          type: number
          format: double
        provinceCode:
          type: string
        cityId:
          type: string
        sources:
          type: object
          additionalProperties: true
    Place:
      type: object
      required:
      - id
      - lat
      - lng
      properties:
        id:
          type: string
        name:
          type: string
        basicCategory:
          type: string
        operatingStatus:
          type: string
        confidence:
          type: number
          format: double
        lat:
          type: number
          format: double
        lng:
          type: number
          format: double
        distanceMeters:
          type: number
          format: double
        provinceCode:
          type: string
        cityId:
          type: string
        websites: {}
        socials: {}
        phones: {}
        emails: {}
        addresses: {}
        taxonomy: {}
        sources: {}
...
    RoadSegment:
      type: object
      required:
      - id
      - subtype
      properties:
        id:
          type: string
        name:
          type: string
        subtype:
          type: string
        class:
          type: string
        subclass:
          type: string
        provinceCode:
          type: string
        distanceMeters:
          type: number
        lengthMeters:
          type: number
        geometry:
          $ref: '#/components/schemas/GeoJSONGeometry'
        speedLimits: {}
        accessRestrictions: {}
        roadSurface: {}
        sources: {}
...
    AddressValidationRequest:
      type: object
      additionalProperties: false
      required:
      - lat
      - lng
      properties:
        lat:
          type: number
          format: double
        lng:
          type: number
          format: double
        provinceCode:
          type: string
        cityId:
          type: string
    AddressValidationResult:
      type: object
      required:
      - countryCode
      - countryName
      - insideCountry
      properties:
        countryCode:
          type: string
        countryName:
          type: string
        insideCountry:
          type: boolean
        provinceCode:
          type: string
        provinceName:
          type: string
        provinceMatches:
          type: boolean
        cityId:
          type: string
        cityName:
          type: string
        cityMatches:
          type: boolean
        nearestPlaces:
          type: array
          items:
            $ref: '#/components/schemas/Place'
        nearestAddresses:
          type: array
          items:
            $ref: '#/components/schemas/Address'
        nearestRoads:
          type: array
          items:
            $ref: '#/components/schemas/RoadSegment'
    GeoJSONGeometry:
      type: object
      required:
      - type
      - coordinates
      properties:
        type:
          type: string
          enum:
          - Point
          - MultiPoint
          - LineString
          - MultiLineString
          - Polygon
          - MultiPolygon
        coordinates: {}
...
    AddressValidationResponse:
      type: object
      required:
      - valid
      - data
      properties:
        valid:
          type: boolean
          example: true
        data:
          $ref: '#/components/schemas/AddressValidationResult'
        message:
          type: string
...